# Faster check for any default scope with all_queries