@zzak 2021-06-02
So I was just curious and it seems `default_scopes?` was deprecated in #10113 and removed in 4.1? What I haven't figured out yet was when it was added back, and why 🤔