@nvasilevski 2021-06-02
I think it was added as part of the `all_queries` scoping feature here - https://github.com/rails/rails/pull/40805/files#diff-f63f081b640fcc6bc7c782ac997566b88a440dba9e06e434db99de94a9d2c8f5R58