@eileencodes 2021-06-03
Thanks for the change and benchmarks! I don't think we default scope stacks are usually that large, but I'd rather it be faster and easier to read regardless.